requirements.txt
scapy
pandas
scikit-learn
flask
joblib
pyyaml
tqdm
python-magic
requests
pytest
